requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==40.0.2
daphne==4.0.0
Django
django-cleanup==7.0.0
django-crispy-forms==2.0
django-discover-runner==1.0
django-role-permissions==3.1.1
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Markdown==3.4.3
msgpack==1.0.5
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
regex==2023.5.5
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
zope.interface==6.0
